🚚 api: move certification attestation model to
certfication course src directory
