+++ api/src/certification/course/infrastructure/repositories/certificate-repository.js
@@ -1,13 +1,9 @@
 import _ from 'lodash';
 import { knex } from '../../../../../db/knex-database-connection.js';
-import {
-  CertificationAttestation,
-  ResultCompetenceTree,
-  CompetenceMark,
-  AssessmentResult,
-} from '../../../../../lib/domain/models/index.js';
+import { ResultCompetenceTree, CompetenceMark, AssessmentResult } from '../../../../../lib/domain/models/index.js';
+import { CertificationAttestation } from '../../domain/models/CertificationAttestation.js';
 import * as competenceTreeRepository from '../../../../../lib/infrastructure/repositories/competence-tree-repository.js';
-import { NotFoundError } from '../../../../../lib/domain/errors.js';
+import { NotFoundError } from '../../../../shared/domain/errors.js';
 import { CertifiedBadge } from '../../../../../lib/domain/read-models/CertifiedBadge.js';
 
 const findByDivisionForScoIsManagingStudentsOrganization = async function ({ organizationId, division }) {
