+++ api/tests/certification/course/unit/domain/models/CertificationAttestation_test.js
@@ -1,4 +1,4 @@
-import { expect, domainBuilder } from '../../../test-helper.js';
+import { expect, domainBuilder } from '../../../../../test-helper.js';
 
 describe('Unit | Domain | Models | CertificationAttestation', function () {
   context('#setResultCompetenceTree', function () {
